fix: move funcao de treino para fora do arquivo
